merge Master

See merge request octopus/autoexec!8

+++ lib/RunNode.py
@@ -75,6 +75,7 @@ class RunNode:
 
         self.statusPath = '{}/{}-{}.json'.format(self.statusPhaseDir, node['host'], node['port'])
 
+        self.outputRoot = self.runPath + '/output'
         self.outputPathPrefix = '{}/output/{}-{}'.format(self.runPath, node['host'], node['port'])
         self.opOutputPathPrefix = '{}/output-op/{}-{}'.format(self.runPath, node['host'], node['port'])
         self.outputPath = self.outputPathPrefix + '.json'
@@ -391,9 +392,13 @@ class RunNode:
 
         cmdline = 'exec {}'.format(orgCmdLine)
         environment = {}
+        environment['OUTPUT_ROOT_PATH'] = self.outputRoot
         environment['OUTPUT_PATH'] = self._getOpOutputPath(op)
         environment['PATH'] = '{}:{}'.format(op.pluginParentPath, os.environ['PATH'])
         environment['PERLLIB'] = '{}/lib:{}'.format(op.pluginParentPath, os.environ['PERLLIB'])
+        environment['AUTOEXEC_JOBID'] = self.context.jobId
+        environment['AUTOEXEC_NODE'] = json.dumps(self.node)
+        environment['AUTOEXEC_NODES_PATH'] = self.context.phases[phaseName].nodesFilePath
 
         scriptFile = None
         if op.isScript == 1:
@@ -444,9 +449,13 @@ class RunNode:
 
         cmdline = 'exec {} --node \'{}\''.format(orgCmdLine, json.dumps(self.node))
         environment = {}
+        environment['OUTPUT_ROOT_PATH'] = self.outputRoot
         environment['OUTPUT_PATH'] = self._getOpOutputPath(op)
         environment['PATH'] = '{}:{}'.format(op.pluginParentPath, os.environ['PATH'])
         environment['PERLLIB'] = '{}/lib:{}'.format(op.pluginParentPath, os.environ['PERLLIB'])
+        environment['AUTOEXEC_JOBID'] = self.context.jobId
+        environment['AUTOEXEC_NODE'] = json.dumps(self.node)
+        environment['AUTOEXEC_NODES_PATH'] = self.context.phases[phaseName].nodesFilePath
 
         scriptFile = None
         if op.isScript == 1:
@@ -524,10 +533,11 @@ class RunNode:
                             self.logHandle.write("ERROR: Download output failed.\n")
                             ret = 2
                     try:
-                        if tagent.agentOsType == 'windows':
-                            tagent.execCmd(self.username, "rd /s /q {}".format(remotePath), env=runEnv)
-                        else:
-                            tagent.execCmd(self.username, "rm -rf {}".format(remotePath), env=runEnv)
+                        if ret != 0 and self.context.devMode:
+                            if tagent.agentOsType == 'windows':
+                                tagent.execCmd(self.username, "rd /s /q {}".format(remotePath), env=runEnv)
+                            else:
+                                tagent.execCmd(self.username, "rm -rf {}".format(remotePath), env=runEnv)
                     except Exception as ex:
                         self.logHandle.write('ERROR: Remote remove directory {} failed {}\n'.format(remotePath, ex))
             except Exception as ex:
@@ -646,7 +656,8 @@ class RunNode:
                             ret = 2
 
                     try:
-                        ssh.exec_command("rm -rf {} || rd /s /q {}".format(remotePath, remotePath))
+                        if ret != 0 and self.context.devMode:
+                            ssh.exec_command("rm -rf {} || rd /s /q {}".format(remotePath, remotePath))
                     except Exception as ex:
                         self.logHandle.write("ERROR: Remove remote directory {} failed {}\n".format(remotePath, ex))
